refactor(driver): share storage helper forwarders
